Pulled common tests to separate file.
