+++ handler/access_log/access_log_test.go
@@ -5,11 +5,9 @@ package access_log
 import (
 	"bytes"
 	"encoding/base64"
-	"errors"
 	"log/slog"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"regexp"
 	"testing"
 
@@ -104,101 +102,5 @@ func TestAccessLoggingUser(t *testing.T) {
 
 	if !userMatch.Match(logBuf.Bytes()) {
 		t.Errorf("user not logged correctly: %v", logBuf.String())
-	}
-}
-
-func TestAccessLoggingNil(t *testing.T) {
-	var handler *Handler
-
-	req := httptest.NewRequest("GET", "/", nil)
-	req.Header.Add("Authorization", "Basic "+base64.StdEncoding.EncodeToString([]byte("testuser:testpass")))
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Result().StatusCode != http.StatusInternalServerError {
-		t.Errorf("expected %v but got %v", http.StatusInternalServerError, rec.Result().StatusCode)
-	}
-}
-
-//
-// Generic Options
-//
-
-func TestOptionError(t *testing.T) {
-	errOpt := func(h *Handler) error {
-		return errors.New("testerror")
-	}
-
-	_, err := New(errOpt)
-
-	if err == nil {
-		t.Errorf("expected middleware creation to fail")
-	}
-}
-
-func TestOptionNil(t *testing.T) {
-	_, err := New(nil)
-
-	if err == nil {
-		t.Errorf("expected middleware creation to fail")
-	}
-}
-
-func TestAccessLoggingNextNil(t *testing.T) {
-	h := util.Must(New(WithLogLevel(slog.LevelDebug)))(nil)
-
-	if h != nil {
-		t.Errorf("expected handler to be nil")
-	}
-}
-
-//
-// WithLevel
-//
-
-func TestOptionWithLevel(t *testing.T) {
-	h := util.Must(New(WithLogLevel(slog.LevelDebug)))(http.HandlerFunc(util.DummyHandler))
-
-	if h.(*Handler).LogLevel() != slog.LevelDebug {
-		t.Errorf("wanted loglevel debug not set")
-	}
-}
-
-func TestOptionWithLevelOnNil(t *testing.T) {
-	err := WithLogLevel(slog.LevelDebug)(nil)
-
-	if err == nil {
-		t.Errorf("expted error on configuring nil handler")
-	}
-}
-
-//
-// WithLogger
-//
-
-func TestOptionWithLogger(t *testing.T) {
-	l := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	h := util.Must(New(WithLogger(l)))(http.HandlerFunc(util.DummyHandler))
-
-	if h.(*Handler).Log() != l {
-		t.Errorf("logger not set correctly")
-	}
-}
-
-func TestOptionWithLoggerOnNil(t *testing.T) {
-	err := WithLogger(slog.Default())(nil)
-
-	if err == nil {
-		t.Errorf("expted error on configuring nil handler")
-	}
-}
-
-func TestOptionWithNilLogger(t *testing.T) {
-	var l *slog.Logger = nil
-	_, hErr := New(WithLogger(l))
-
-	if hErr == nil {
-		t.Errorf("expected error on configuration with nil logger")
 	}
 }
