+++ handler/correlation/correlation_test.go
@@ -3,10 +3,8 @@
 package correlation
 
 import (
-	"log/slog"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"testing"
 
 	"github.com/AlphaOne1/midgard/util"
@@ -60,55 +58,5 @@ func TestCorrelationSuppliedID(t *testing.T) {
 
 	if rec.Header().Get("X-Correlation-iD") != "setOutside" {
 		t.Errorf("X-Correlation-ID header not set correctly in response")
-	}
-}
-
-//
-// WithLevel
-//
-
-func TestOptionWithLevel(t *testing.T) {
-	h := util.Must(New(WithLogLevel(slog.LevelDebug)))(http.HandlerFunc(util.DummyHandler))
-
-	if h.(*Handler).LogLevel() != slog.LevelDebug {
-		t.Errorf("wanted loglevel debug not set")
-	}
-}
-
-func TestOptionWithLevelOnNil(t *testing.T) {
-	err := WithLogLevel(slog.LevelDebug)(nil)
-
-	if err == nil {
-		t.Errorf("expted error on configuring nil handler")
-	}
-}
-
-//
-// WithLogger
-//
-
-func TestOptionWithLogger(t *testing.T) {
-	l := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	h := util.Must(New(WithLogger(l)))(http.HandlerFunc(util.DummyHandler))
-
-	if h.(*Handler).Log() != l {
-		t.Errorf("logger not set correctly")
-	}
-}
-
-func TestOptionWithLoggerOnNil(t *testing.T) {
-	err := WithLogger(slog.Default())(nil)
-
-	if err == nil {
-		t.Errorf("expted error on configuring nil handler")
-	}
-}
-
-func TestOptionWithNilLogger(t *testing.T) {
-	var l *slog.Logger = nil
-	_, hErr := New(WithLogger(l))
-
-	if hErr == nil {
-		t.Errorf("expected error on configuration with nil logger")
 	}
 }
